requirements.txt
fastapi
uvicorn[standard]
docker
requests
mysql-connector-python
python-dotenv
cryptography
